fixed bug in run_strainest.sh for synthetic

+++ examples/synthetic/scripts/read_depths/run_strainest.sh
@@ -67,7 +67,18 @@ rm ${sam_file}
 rm ${bam_file}
 rm ${sorted_bam_file}
 
-mv abund.txt abund_${time_point}.txt
-mv info.txt info_${time_point}.txt
-mv counts.txt counts_${time_point}.txt
-mv max_ident.txt max_ident_${time_point}.txt
+if [ -f abund.txt ]; then
+	mv abund.txt abund_${time_point}.txt
+fi
+
+if [ -f info.txt ]; then
+	mv info.txt info_${time_point}.txt
+fi
+
+if [ -f counts.txt ]; then
+	mv counts.txt counts_${time_point}.txt
+fi
+
+if [ -f max_ident.txt ]; then
+	mv max_ident.txt max_ident_${time_point}.txt
+fi
